// quarkus-astro-app/src/main/webui/src/lib/api/model/categoryCreateDTO.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Task Manager API
 * RESTful API for Task Manager application demonstrating Islands Architecture
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryCreateDTO {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 7
   * @pattern ^#[0-9A-Fa-f]{6}$
   */
  colorCode: string;
}